swap the module relationship between auction future and worker task
